+++ src/Fauna/herbivore_cohort.cpp
@@ -14,10 +14,11 @@
 
 using namespace Fauna;
 
-HerbivoreCohort::HerbivoreCohort(
-    const int age_days, const double body_condition,
-    std::shared_ptr<const Hft> hft, const Sex sex, const double ind_per_km2,
-    const ForageEnergyContent& forage_gross_energy)
+HerbivoreCohort::HerbivoreCohort(const int age_days,
+                                 const double body_condition,
+                                 std::shared_ptr<const Hft> hft, const Sex sex,
+                                 const double ind_per_km2,
+                                 const ForageEnergyContent& forage_gross_energy)
     : HerbivoreBase(age_days, body_condition, hft, sex, forage_gross_energy),
       ind_per_km2(ind_per_km2) {
   if (ind_per_km2 < 0.0)
@@ -26,9 +27,9 @@ HerbivoreCohort::HerbivoreCohort(
         "ind_per_km2 <0.0");
 }
 
-HerbivoreCohort::HerbivoreCohort(
-    std::shared_ptr<const Hft> hft, const Sex sex, const double ind_per_km2,
-    const ForageEnergyContent& forage_gross_energy)
+HerbivoreCohort::HerbivoreCohort(std::shared_ptr<const Hft> hft, const Sex sex,
+                                 const double ind_per_km2,
+                                 const ForageEnergyContent& forage_gross_energy)
     : HerbivoreBase(hft, sex, forage_gross_energy), ind_per_km2(ind_per_km2) {
   if (ind_per_km2 < 0.0)
     throw std::invalid_argument(