+++ src/Fauna/net_energy_models.h
@@ -46,15 +46,15 @@ namespace Fauna {
  * \see \ref NetEnergyModel
  */
 inline ForageEnergyContent get_net_energy_from_gross_energy(
-    const ForageEnergyContent& ge_content,
-    const Digestibility& digestibility,
-    const double me_coefficient,
-    const double k_maintenance){
+    const ForageEnergyContent& ge_content, const Digestibility& digestibility,
+    const double me_coefficient, const double k_maintenance) {
   if (me_coefficient <= 0.0 || me_coefficient >= 1.0)
-    throw std::invalid_argument("Fauna::get_net_energy_from_gross_energy(): "
+    throw std::invalid_argument(
+        "Fauna::get_net_energy_from_gross_energy(): "
         "Parameter `me_coefficient` is not in interval (0,1).");
   if (k_maintenance <= 0.0 || k_maintenance >= 1.0)
-    throw std::invalid_argument("Fauna::get_net_energy_from_gross_energy(): "
+    throw std::invalid_argument(
+        "Fauna::get_net_energy_from_gross_energy(): "
         "Parameter `k_maintenance` is not in interval (0,1).");
   return ge_content * digestibility * (me_coefficient * k_maintenance);
 }